Fixed compile warning
<==== Mercurial data ====>
Branch: dev_3.0.0

+++ common_libs/nx_network/src/nx/network/test_support/simple_socket_test_helper.h
@@ -394,7 +394,7 @@ void socketTransferAsync(
                     return serverResults.push(code);
 
                 EXPECT_STREQ(serverBuffer.data(), testMessage.data());
-                if (size < testMessage.size())
+                if (size < (size_t)testMessage.size())
                     return serverResults.push(SystemError::connectionReset);
 
                 serverBuffer.resize(0);
